requirements.txt
django==2.1
djangorestframework==3.8.2
flake8==3.5.0
Pillow==5.2.0
django-rest-auth==0.9.3
coverage==4.5.1
django-rest-swagger==2.2.0
